app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
requests
langchain
python-dotenv
PyMuPDF
